port functional and consistency tests
